storage: widen get_resolved flags to a reserved u24, condense docs
flags becomes a reserved 24-bit field carried in a u32, so future options can be
added without another wire-format change. Transmitted as 3 bytes so that key_type +
flags pack into one 4-byte tail, keeping the request 52 bytes -- a multiple of 4.

Unknown bits are rejected rather than ignored, so a caller can tell whether an
option took effect; the client also refuses flags that would not fit on the wire
instead of silently truncating them.

Drops the ALSO_SUBFRAGMENTS bit definition: pushing referenced subfragments needs a
streaming response, which the QUIC protocol does not have (one response per
command_id, and a second frame is fatal), so no bit is defined yet.

+++ lore-transport/src/session.rs
@@ -230,15 +230,14 @@ impl StorageSession {
         storage.mutable_load(session_id, key, key_type).await
     }
 
-    /// Resolve a mutable key and read the immutable blob it points at in ONE round trip,
-    /// instead of `mutable_load` followed by `get`. Returns `(resolved_hash, fragment,
-    /// payload)`. `flags` is a bitmask of `GetResolvedFlags` (0 for default behaviour).
+    /// `mutable_load` + `get` in one round trip. Returns `(resolved_hash, fragment, payload)`.
+    /// `flags` is a `get_resolved_flags` bitmask; 0 for default behaviour.
     pub async fn get_resolved(
         &self,
         key: &Hash,
         key_type: KeyType,
         context: &Context,
-        flags: u8,
+        flags: u32,
     ) -> Result<(Hash, Fragment, Bytes), ProtocolError> {
         let (storage, session_id) = self.ensure().await?;
         storage
